进行简单等代码审查，修复split index error等bug，merge size缺少的bug尚未修复

+++ murmurhash.c
@@ -8,6 +8,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <stdint.h>
+#include <string.h>
 #include "murmurhash.h"
 
 //demo:
@@ -93,6 +94,6 @@ char *murmurhash_str(const char *origin_str)
 	char *mmh_str;
 	mmh_str  = (char *)calloc(32, sizeof(char));
 	mmh = murmurhash(origin_str, (uint32_t )strlen(origin_str), 1024);
-	sprintf(mmh_str, "%ld", mmh);
+	sprintf(mmh_str, "%ld", (long)mmh);
 	return mmh_str;
 }
